@@ -15,20 +15,21 @@
     "clean": "rimraf wwwroot/dist clientapp/dist"
   },
   "dependencies": {
-    "@angular/animations": "^4.3.0",
-    "@angular/common": "^4.3.0",
-    "@angular/compiler": "^4.3.0",
-    "@angular/compiler-cli": "^4.3.0",
-    "@angular/core": "^4.3.0",
-    "@angular/forms": "^4.3.0",
-    "@angular/http": "^4.3.0",
-    "@angular/platform-browser": "^4.3.0",
-    "@angular/platform-browser-dynamic": "^4.3.0",
-    "@angular/platform-server": "^4.3.0",
-    "@angular/router": "^4.3.0",
-    "@nguniversal/aspnetcore-engine": "^1.0.0-beta.2",
-    "@ngx-translate/core": "^6.0.1",
-    "@ngx-translate/http-loader": "0.0.3",
+    "@angular/animations": "^5.0.0",
+    "@angular/common": "^5.0.0",
+    "@angular/compiler": "^5.0.0",
+    "@angular/compiler-cli": "^5.0.0",
+    "@angular/core": "^5.0.0",
+    "@angular/forms": "^5.0.0",
+    "@angular/http": "^5.0.0",
+    "@angular/platform-browser": "^5.0.0",
+    "@angular/platform-browser-dynamic": "^5.0.0",
+    "@angular/platform-server": "^5.0.0",
+    "@angular/router": "^5.0.0",
+    "@nguniversal/aspnetcore-engine": "^5.0.0-beta.5",
+    "@nguniversal/common": "^5.0.0-beta.5",
+    "@ngx-translate/core": "^8.0.0",
+    "@ngx-translate/http-loader": "^2.0.0",
     "@types/node": "^7.0.12",
     "angular2-router-loader": "^0.3.5",
     "angular2-template-loader": "^0.6.2",
@@ -49,7 +50,7 @@
     "jquery": "^2.2.1",
     "json-loader": "^0.5.4",
     "moment": "2.18.1",
-    "ngx-bootstrap": "2.0.0-beta.3",
+    "ngx-bootstrap": "2.0.0-beta.6",
     "node-sass": "^4.5.2",
     "preboot": "^5.0.0",
     "raw-loader": "^0.5.1",
@@ -58,7 +59,7 @@
     "sass-loader": "^6.0.6",
     "style-loader": "^0.18.2",
     "to-string-loader": "^1.1.5",
-    "typescript": "2.5.2",
+    "typescript": "2.5.3",
     "url-loader": "^0.5.7",
     "webpack": "^3.6.0",
     "webpack-hot-middleware": "^2.19.1",
@@ -66,8 +67,8 @@
     "zone.js": "^0.8.17"
   },
   "devDependencies": {
-    "@angular/cli": "^1.3.2",
-    "@ngtools/webpack": "^1.3.0",
+    "@angular/cli": "^1.5.0",
+    "@ngtools/webpack": "^1.8.0",
     "@types/chai": "^3.4.34",
     "@types/jasmine": "^2.5.37",
     "chai": "^3.5.0",
