@@ -3,12 +3,10 @@ import { BrowserModule } from '@angular/platform-browser';
 import { APP_BASE_HREF } from '@angular/common';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
-import { ORIGIN_URL } from './shared/constants/baseurl.constants';
+import { ORIGIN_URL, REQUEST } from '@nguniversal/aspnetcore-engine';
 import { AppModuleShared } from './app.module';
 import { AppComponent } from './app.component';
-import { REQUEST } from './shared/constants/request';
-import { BrowserTransferStateModule } from '../modules/transfer-state/browser-transfer-state.module';
-
+import { BrowserTransferStateModule } from '@angular/platform-browser';
 import { BrowserPrebootModule } from 'preboot/browser';
 
 export function getOriginUrl() {
@@ -23,12 +21,8 @@ export function getRequest() {
 @NgModule({
     bootstrap: [AppComponent],
     imports: [
-        BrowserModule.withServerTransition({
-            appId: 'my-app-id' // make sure this matches with your Server NgModule
-        }),
         BrowserPrebootModule.replayEvents(),
         BrowserAnimationsModule,
-        BrowserTransferStateModule,
 
         // Our Common AppModule
         AppModuleShared
