@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 
 @Component({
-    selector: 'not-found',
+    selector: 'app-not-found',
     templateUrl: './not-found.component.html'
 })
 export class NotFoundComponent implements OnInit {
