@@ -54,7 +54,7 @@ namespace AspCoreServer
       // Register the Swagger generator, defining one or more Swagger documents
       services.AddSwaggerGen(c =>
       {
-        c.SwaggerDoc("v1", new Info { Title = "Angular 4.0 Universal & ASP.NET Core advanced starter-kit web API", Version = "v1" });
+        c.SwaggerDoc("v1", new Info { Title = "Angular 5.0 Universal & ASP.NET Core advanced starter-kit web API", Version = "v1" });
       });
     }
 
@@ -74,7 +74,7 @@ namespace AspCoreServer
         app.UseWebpackDevMiddleware(new WebpackDevMiddlewareOptions
         {
           HotModuleReplacement = true,
-          HotModuleReplacementEndpoint = "/dist/__webpack_hmr"
+          HotModuleReplacementEndpoint = "/dist/"
         });
         app.UseSwagger();
         app.UseSwaggerUI(c =>
